detectionsystemrpi/requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
firebase-admin==2.16.0
google-api-core==1.8.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-cloud-vision==0.36.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
httplib2==0.12.1
idna==2.8
msgpack==0.6.1
picamera==1.13
pkg-resources==0.0.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-firebase==1.2
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
